trackproject/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
drf-spectacular==0.26.4
psycopg2-binary==2.9.7
gunicorn==21.2.0
boto3==1.28.49
